c4isr/backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
python-dotenv==1.0.1
cachetools==5.5.0